+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * feat: Add rule processing logging
 * feat: Enhance helpers functionalities (especially related to `type` handling)
 * feat: Loop over `update_schema` for continuos optimisation
+* feat: Add tests against JSON-Test-Schema-Suite to preserve and ensure correctness
 
 ## 0.1.0 (2020-06-07)
 
@@ -18,5 +19,3 @@ Initial release.
 * Handling of extraneous keys with respect to `type` keyword
 * Suppression of empty `additionalProperties` and `required`
 * Suppression of impossible schemas caused by `min`/`max` related keywords
-
-